refactor: expert 관련 도메인 패키지 추가
